particles refactor, use strict, and google analyics added
