+++ src/app/actors/PhysicsActor.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var properties = require('../properties');
 
 /**
