+++ src/app/actors/PowerStation.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var PhysicsActor = require('./PhysicsActor');
 
 /**
