+++ src/app/actors/strategies/FiringStrategy.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var properties = require('../../properties');
 
 /**
