+++ src/app/actors/strategies/ForwardFiring.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var FiringStrategy = require('./FiringStrategy');
 var _ = require('lodash');
