+++ src/app/environment/particles/manager.js
@@ -3,7 +3,8 @@
  * Manager for particle effects,
  * currently houses the swirl effect uses when the player jumps into a level
  *
- * @class particles
+ * @class manager
+ * @namespace environment.particles
  * @type {{group: null, manager: null, emitter: null, init: Function, create: Function, startSwirl: Function}}
  * @static
  */
@@ -20,8 +21,6 @@ module.exports = {
    * @type {Phaser.ParticleStorm.Emitter}
    */
   emitter: null,
-
-  noRefuel: true,
 
   /**
    * @method init
